Move caret code to its own package
Subpackage of websockets package
