+++ websockets/caret/caret_add_test.go
@@ -1,4 +1,4 @@
-package websockets
+package caret
 
 var addTests = []struct {
 	Name            string
