+++ websockets/caret/caret_replace_test.go
@@ -1,4 +1,4 @@
-package websockets
+package caret
 
 var replaceTests = []struct {
 	Name            string
